requirements.txt
annotated-types==0.7.0
anyio==4.9.0
baostock==0.8.9
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
distro==1.9.0
fastapi==0.115.12
filetype==1.2.0
google-ai-generativelanguage==0.6.18
google-api-core==2.24.2
google-auth==2.40.2
googleapis-common-protos==1.70.0
grpcio==1.73.0rc1
grpcio-status==1.73.0rc1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.3.62
langchain-google-genai==2.1.4
langchain-mcp-adapters==0.1.1
langchain-openai==0.3.18
langgraph==0.4.7
langgraph-checkpoint==2.0.26
langgraph-prebuilt==0.2.1
langgraph-sdk==0.1.70
langsmith==0.3.42
markdown-it-py==3.0.0
mcp==1.8.0
mdurl==0.1.2
nest-asyncio==1.6.0
numpy==2.2.6
openai==1.78.0
orjson==3.10.18
ormsgpack==1.10.0
packaging==24.2
pandas==2.2.3
proto-plus==1.26.1
protobuf==6.31.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
rsa==4.9.1
setuptools==78.1.1
six==1.17.0
sniffio==1.3.1
socksio==1.0.0
sse-starlette==2.3.4
starlette==0.46.2
tabulate==0.9.0
tenacity==9.1.2
tiktoken==0.9.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
websockets==15.0.1
wheel==0.45.1
xxhash==3.5.0
zstandard==0.23.0
